rfactor

+++ convex/board.ts
@@ -1,7 +1,5 @@
 import { v } from "convex/values";
-
 import { mutation, query } from "./_generated/server";
-
 const images = [
     "/placeholders/1.svg",
     "/placeholders/2.svg",
@@ -22,15 +20,11 @@ export const create = mutation({
     },
     handler: async (ctx, args) => {
         const identity = await ctx.auth.getUserIdentity();
-
         if (!identity) {
             throw new Error("Unauthorized");
         }
 
         const randomImage = images[Math.floor(Math.random() * images.length)];
-
-        console.log(randomImage, "TEST")
-
         const board = await ctx.db.insert("boards", {
             title: args.title,
             orgId: args.orgId,
@@ -44,22 +38,19 @@ export const create = mutation({
 });
 
 export const remove = mutation({
-    args: { id: v.id("boards") },
+    args: {
+        id: v.id("boards"),
+    },
     handler: async (ctx, args) => {
         const identity = await ctx.auth.getUserIdentity();
-
         if (!identity) {
             throw new Error("Unauthorized");
         }
-
         const userId = identity.subject;
-
         const existingFavorite = await ctx.db
             .query("userFavorites")
             .withIndex("by_user_board", (q) =>
-                q
-                    .eq("userId", userId)
-                    .eq("boardId", args.id)
+                q.eq("userId", userId).eq("boardId", args.id)
             )
             .unique();
 
@@ -67,48 +58,46 @@ export const remove = mutation({
             await ctx.db.delete(existingFavorite._id);
         }
 
+        const board = await ctx.db.get(args.id);
+        if (board?.authorId !== userId) throw new Error("Unauthorized");
+
         await ctx.db.delete(args.id);
     },
 });
 
 export const update = mutation({
-    args: { id: v.id("boards"), title: v.string() },
+    args: {
+        id: v.id("boards"),
+        title: v.string(),
+    },
     handler: async (ctx, args) => {
         const identity = await ctx.auth.getUserIdentity();
-
         if (!identity) {
             throw new Error("Unauthorized");
         }
 
         const title = args.title.trim();
+        if (!title) throw new Error("Title is required");
 
-        if (!title) {
-            throw new Error("Title is required");
-        }
+        if (title.length > 60)
+            throw new Error("Title cannot be longer than 60 characters");
 
-        if (title.length > 60) {
-            throw new Error("Title cannot be longer than 60 characters")
-        }
-
-        const board = await ctx.db.patch(args.id, {
-            title: args.title,
-        });
-
-        return board;
+        return await ctx.db.patch(args.id, { title: args.title });
     },
 });
 
 export const favorite = mutation({
-    args: { id: v.id("boards"), orgId: v.string() },
+    args: {
+        id: v.id("boards"),
+        orgId: v.string(),
+    },
     handler: async (ctx, args) => {
         const identity = await ctx.auth.getUserIdentity();
-
         if (!identity) {
             throw new Error("Unauthorized");
         }
 
         const board = await ctx.db.get(args.id);
-
         if (!board) {
             throw new Error("Board not found");
         }
@@ -118,16 +107,11 @@ export const favorite = mutation({
         const existingFavorite = await ctx.db
             .query("userFavorites")
             .withIndex("by_user_board", (q) =>
-                q
-                    .eq("userId", userId)
-                    .eq("boardId", board._id)
+                q.eq("userId", userId).eq("boardId", board._id)
             )
             .unique();
 
-        if (existingFavorite) {
-            throw new Error("Board already favorited");
-        }
-
+        if (existingFavorite) throw new Error("Board already favorited");
         await ctx.db.insert("userFavorites", {
             userId,
             boardId: board._id,
@@ -138,18 +122,17 @@ export const favorite = mutation({
     },
 });
 
-
 export const unfavorite = mutation({
-    args: { id: v.id("boards") },
+    args: {
+        id: v.id("boards"),
+    },
     handler: async (ctx, args) => {
         const identity = await ctx.auth.getUserIdentity();
-
         if (!identity) {
             throw new Error("Unauthorized");
         }
 
         const board = await ctx.db.get(args.id);
-
         if (!board) {
             throw new Error("Board not found");
         }
@@ -159,15 +142,11 @@ export const unfavorite = mutation({
         const existingFavorite = await ctx.db
             .query("userFavorites")
             .withIndex("by_user_board", (q) =>
-                q
-                    .eq("userId", userId)
-                    .eq("boardId", board._id)
+                q.eq("userId", userId).eq("boardId", board._id)
             )
             .unique();
 
-        if (!existingFavorite) {
-            throw new Error("Favorited board not found");
-        }
+        if (!existingFavorite) throw new Error("Favorited board not found");
 
         await ctx.db.delete(existingFavorite._id);
 
@@ -176,10 +155,23 @@ export const unfavorite = mutation({
 });
 
 export const get = query({
-    args: { id: v.id("boards") },
+    args: {
+        id: v.id("boards"),
+    },
     handler: async (ctx, args) => {
-        const board = ctx.db.get(args.id);
+        return await ctx.db.get(args.id);
+    },
+});
 
-        return board;
+export const getTotalBoardCountOfOrg = query({
+    args: {
+        orgId: v.string(),
+    },
+    handler: async (ctx, args) => {
+        return await ctx.db
+            .query("boards")
+            .withIndex("by_org", (q) => q.eq("orgId", args.orgId))
+            .collect()
+            .then((boards) => boards.length);
     },
 });
